[T] move component
